changed the function argument type

+++ include/task2.h
@@ -1,3 +1,3 @@
 bool checkPrime(unsigned int value);  //проверка числа на простоту
-unsigned int nPrime(unsigned n);  //нахождение n-ого простого числа (в ряду)
-unsigned int nextPrime(unsigned int value);  //нахождение ближайшего следующего простого числа к value
+unsigned long long nPrime(unsigned n);  //нахождение n-ого простого числа (в ряду)
+unsigned long long nextPrime(unsigned long long value);  //нахождение ближайшего следующего простого числа к value
