+++ include/task3.h
@@ -1,3 +1,3 @@
-unsigned long sumPrime(unsigned int hbound);  //сумма всех простых чисел до hbound (не включая его)
+unsigned long long sumPrime(unsigned int hbound);  //сумма всех простых чисел до hbound (не включая его)
 
 //bool checkPrime(unsigned int value);  //проверка числа на простоту
